Connection with database made through Context Manager

+++ src/data_management/tf_idf/tf_idf_op.py
@@ -31,7 +31,7 @@ class TfIDfOperator:
         return norm_vec
 
     @staticmethod
-    def get_cos(tf_idf_matrix: ndarray, docs: List[str], query_vec: ndarray) -> List[List[float64 | str]]:
+    def get_cos(tf_idf_matrix: ndarray, docs: List[str], query_vec: ndarray) -> List[List[float | str]]:
         '''
         This function will get the tf_idf_matrix at the space IRn and a list
         of available documents that were used inside the matrix in the same 
@@ -47,7 +47,7 @@ class TfIDfOperator:
         return >> list with list with cossine value of the angle between the query vector and the vector at each row aligned with its document string.
         '''
 
-        organized_cos: List[List[float64 | str]] = [
+        organized_cos: List[List[float | str]] = [
             [
                 TfIDfOperator.get_inner_product(
                 TfIDfOperator.vec_normalizer(query_vec),
